core: add docs for imgstry classes and types
