+++ source/decorators/bind-utility.decorator.ts
@@ -1,3 +1,11 @@
+/**
+ * Binds modules to a property.
+ *
+ * @export
+ * @param {(...Record<string, Function | Object>[])} utilities
+ * @returns
+ * @ignore
+ */
 export function BindUtility(...utilities: Record<string, Function | Object>[]) {
   return function (target: any, propertyKey: string) {
     const utility: Record<string, Function | Object> = {};
